requirements.txt
# Django framework
Django>=4.2,<5.0
django-environ>=0.9.0

# Database
psycopg2-binary>=2.9.0

# Data analysis and visualization
pandas>=1.3.0
plotly>=5.0.0

# Environment variables
python-dotenv>=0.19.0

# Web server (for production)
gunicorn>=20.1.0

# Optional: Flask (if you still need it for other scripts)
flask>=2.0.0    